Remove warnings when compiling in browser + speedup
Previously the typescript libraries were provided via tsvfs.createDefaultMapFromCDN
which includes additional definition files that we do not use, including dom.d.ts

This produces irrelevant warnings when compiling the typescript source files in the browser.
We now generate a lib_dts.ts, similar to behavior_dts.ts, which contains the relevant library
files that are used during compilation.

An added benefit is that compilation feels faster than before.
