+++ .gitignore
@@ -6,6 +6,7 @@ website/monaco-editor-0.45.0
 # Generated files
 behavior.d.ts
 behavior_dts.ts
+lib_dts.ts
 methods.ts
 decompile/dsinstr.ts
